brainstorm/datasets/initial_datasets.py
import pickle
from textwrap import dedent
from typing import Any

import pandas as pd
from google.api_core.exceptions import NotFound

from brainstorm.common.utils.io import cloud_storage_upload_string, cloud_storage_download_string, archive_and_upload, download_and_unpack
from brainstorm.db.models.datasets import DatasetDefinition
from brainstorm.core.config import settings

import datasets as hf_datasets


class Dataset:
    def __init__(self, dataset_definition: DatasetDefinition):
        self.dataset_definition = dataset_definition
        self.parsed_data_path_prefix = f'parsed_data/{self.dataset_definition.dataset_id}'

        self.data = self.load_data()

        if dataset_definition.sample_size:
            for split_key in self.data.keys():
                self.data[split_key] = self.data[split_key].select(range(dataset_definition.sample_size))

    def keys(self):
        return self.data.keys()

    def __getitem__(self, key: str) -> hf_datasets.Dataset:
        return self.data[key]

    def download_raw_data(self) -> Any:
        """
        Must handle downloading in whatever format from whatever source and save raw data to gs in whatever original format
        we have received it
        """
        raise NotImplementedError

    def download_parsed_data(self) -> hf_datasets.DatasetDict | None:
        try:
            download_and_unpack(settings.DATA_BUCKET,
                                gcs_prefix=self.parsed_data_path_prefix,
                                extract_to_path=self.parsed_data_path_prefix)
        except NotFound:
            return

        parsed_data = hf_datasets.load_from_disk(self.parsed_data_path_prefix)
        return parsed_data

    def parse(self, raw_data: Any) -> hf_datasets.DatasetDict:
        """
        Handles the data in whatever format and parses it into HFDataset with required and optional fields

        Required fields should include
        - Input - a final input that must be passed to a model, including prompt and all that
        - Output - a final output against which the model's output will be compared by scorer
        """
        raise NotImplementedError

    def load_data(self) -> hf_datasets.DatasetDict:
        parsed_data = self.download_parsed_data()
        if parsed_data is None:
            raw_data = self.download_raw_data()
            parsed_data = self.parse(raw_data=raw_data)

            parsed_data.save_to_disk(self.parsed_data_path_prefix)
            archive_and_upload(
                local_dir_path=self.parsed_data_path_prefix,
                bucket_name=settings.DATA_BUCKET,
                gcs_prefix=self.parsed_data_path_prefix
            )

        return parsed_data


class TruthfulQA(Dataset):
    def download_raw_data(self):
        data = hf_datasets.load_dataset("domenicrosati/TruthfulQA")
        data_prefix = 'raw_data/TruthfulQA'
        data.save_to_disk(data_prefix)
        archive_and_upload(
            local_dir_path=data_prefix,
            bucket_name=settings.DATA_BUCKET,
            gcs_prefix=data_prefix
        )
        return data

    def parse(self, raw_data: Any) -> hf_datasets.DatasetDict:
        # no need to copy since we won't be saving raw data again
        data = raw_data
        for split in data.keys():
            data[split] = data[split].rename_column("Question", "input")
            data[split] = data[split].rename_column('Best Answer', "output")
        return data


class MuSR(Dataset):
    def download_raw_data(self):
        data = hf_datasets.load_dataset("TAUR-Lab/MuSR")
        data_prefix = 'raw_data/MuSR'
        data.save_to_disk(data_prefix)
        archive_and_upload(
            local_dir_path=data_prefix,
            bucket_name=settings.DATA_BUCKET,
            gcs_prefix=data_prefix
        )
        return data

    def parse(self, raw_data: Any) -> hf_datasets.DatasetDict:
        data = raw_data

        def make_input_prompt(x):
            prompt = f"""
            Given a narrative, question and list of possible answer choices below. Don't provide any information, only respond with the correct answer index, for example "0".
            Suppose you have 3 choices: ["Jon", "Sam", "Someone else"] and you think correct answer is "Same", the correct response is "1".
            
            Narrative:
            {x['narrative']}
            
            Question:
            {x['question']}
            
            Possible answer choices:
            {x['choices']}
            """
            return {'input': dedent(prompt)}

        for split in data.keys():
            data[split] = data[split].map(make_input_prompt)
            data[split] = data[split].rename_column('answer_index', "output")

        return data


class BBQ(Dataset):
    def download_raw_data(self):
        data = hf_datasets.load_dataset("Elfsong/BBQ")
        data_prefix = 'raw_data/BBQ'
        data.save_to_disk(data_prefix)
        archive_and_upload(
            local_dir_path=data_prefix,
            bucket_name=settings.DATA_BUCKET,
            gcs_prefix=data_prefix
        )
        return data

    def parse(self, raw_data: Any) -> hf_datasets.DatasetDict:
        data = raw_data

        def make_input_prompt(x):
            prompt = f"""
            Given a narrative, question and list of possible answer choices below. Don't provide any information, only respond with the correct answer index, for example "0".
            Suppose you have 3 choices: ["Jon", "Sam", "Someone else"] and you think correct answer is "Same", the correct response is "1".
            
            Context:
            {x['context']}
            
            Question:
            {x['question']}
            
            Possible answer choices:
            [{x['ans0']}, {x['ans1']}, {x['ans2']}]
            """
            return {'input': dedent(prompt)}

        for split in data.keys():
            data[split] = data[split].map(make_input_prompt)
            data[split] = data[split].rename_column('answer_label', "output")

        return data


DATASETS_MAP = {
    "TruthfulQA": TruthfulQA,
    "MuSR": MuSR,
    "BBQ": BBQ,
}
