pyproject.toml
[tool.poetry]
name = "brainstorm-test-api"
version = "0.1.0"
description = ""
authors = ["Tyler", "Akhat"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
datasets = ">=2.14.0"
openai = ">=1.0.0"
python-dotenv = ">=1.0.0"
pydantic = ">=2.0.0"
pydantic-settings = ">=2.8.1"
requests = ">=2.31.0"
tqdm = ">=4.65.0"
fastapi = ">=0.104.0"
uvicorn = ">=0.23.2"
sqlalchemy = ">=2.0.22"
alembic = ">=1.12.0"
python-multipart = ">=0.0.6"
httpx = ">=0.25.0"
platformdirs = ">=4.3.7"
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
docker = ">=6.1.0"
python-jose = ">=3.3.0"
cryptography = ">=41.0.0"
passlib = ">=1.7.4"
bcrypt = ">=4.0.1"
nltk = ">=3.8.1"
sentence-transformers = ">=2.2.2"
transformers = ">=4.30.0"
torch = ">=2.0.0"
tensorflow = ">=2.13.0"
tensorflow-hub = ">=0.14.0"
bert-score = ">=0.3.13"
nlpaug = ">=1.1.11"
detoxify = ">=0.5.2"
spacy = ">=3.6.0"
stanza = ">=1.5.0"
websockets = "^15.0.1"
psycopg2 = "^2.9.10"
google-auth = "^2.39.0"
google-cloud-storage = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
